refactor(daemon): split server/lifecycle.rs under the 1k-LOC budget
lifecycle.rs (1146 LOC) -> lifecycle.rs (450) + embedded.rs (397, all
EmbeddedDaemon code) + loaders.rs (324, DepGraphSetter + the four *Loader
types + their DaemonServer factory methods). Public paths preserved via
mod.rs re-exports (pub use loaders::DepGraphSetter). Behavior-preserving.
