+++ crates/zccache/src/daemon/server/README.md
@@ -6,7 +6,9 @@ their items so every sibling can `use super::*;` to see the same vocabulary.
 
 Topic-focused submodules:
 
-- `lifecycle.rs` — `DaemonServer::{bind, bind_with_cache_dir}`, accessors, test seams
+- `lifecycle.rs` — `DaemonServer::{bind, bind_with_cache_dir}`, `new_shared_state`, accessors, test seams
+- `embedded.rs` — `EmbeddedDaemon` (in-process host integration): construction, background cache loads, compile entrypoint, flush/shutdown
+- `loaders.rs` — deferred cache-load handles (`DepGraphSetter`, the four `*Loader`s) + the `DaemonServer` factory methods that hand them out (bind-first / load-in-background, #640/#784)
 - `run.rs` — `DaemonServer::run` main loop + watcher pipeline initializer
 - `state.rs` — `SharedState`, the daemon's central state object
 - `cached_artifact.rs` — `CachedArtifact`, `CachedPayload`, payload materialization, legacy `.meta` migration
