+++ crates/zccache/src/daemon/server/lifecycle.rs
@@ -9,16 +9,6 @@ use super::*;
 /// artifact and depfile directories, even within the same process.
 pub(super) static SERVER_INSTANCE: AtomicU64 = AtomicU64::new(0);
 pub(super) static ARTIFACT_PERSIST_TMP_COUNTER: AtomicU64 = AtomicU64::new(0);
-/// zccache#940 — monotonic per-process compile counter for the inner
-/// diagnostic trace. Resets across process restarts by design (the
-/// trace file is process-scoped). Hosts that need durable ids should
-/// cross-correlate by `ts_ns` against their own audit log.
-static INNER_COMPILE_SEQ: AtomicU64 = AtomicU64::new(0);
-
-fn next_inner_compile_id() -> String {
-    let n = INNER_COMPILE_SEQ.fetch_add(1, Ordering::Relaxed);
-    format!("z{n:08x}")
-}
 
 impl DaemonServer {
     /// Create a new daemon server bound to the given endpoint, using the
@@ -53,7 +43,7 @@ impl DaemonServer {
     }
 }
 
-fn new_shared_state(
+pub(super) fn new_shared_state(
     endpoint: &str,
     cache_dir: &crate::core::NormalizedPath,
     backend_identity: running_process::broker::protocol_v2::backend_handle::DaemonProcess,
@@ -238,383 +228,6 @@ fn new_shared_state(
     )
 }
 
-impl EmbeddedDaemon {
-    pub(crate) async fn start(
-        endpoint: String,
-        cache_dir: crate::core::NormalizedPath,
-        runtime_handle: Option<tokio::runtime::Handle>,
-    ) -> Result<Self, crate::ipc::IpcError> {
-        let backend_identity = crate::ipc::current_backend_identity(&endpoint)
-            .map_err(|err| crate::ipc::IpcError::Endpoint(err.to_string()))?;
-        let (state, index_writer_rx) = new_shared_state(&endpoint, &cache_dir, backend_identity);
-
-        let mut daemon = Self {
-            state,
-            index_writer_rx: Some(index_writer_rx),
-            index_writer_handle: Mutex::new(None),
-        };
-        daemon.start_background_tasks(runtime_handle).await;
-        Ok(daemon)
-    }
-
-    async fn start_background_tasks(&mut self, runtime_handle: Option<tokio::runtime::Handle>) {
-        if let Some(rx) = self.index_writer_rx.take() {
-            let store = Arc::clone(&self.state.artifact_store);
-            let shutdown = Arc::clone(&self.state.index_writer_shutdown);
-            let task = run_index_writer(rx, store, shutdown);
-            // zccache#922: when the embedded host supplied a Tokio Handle,
-            // route the persistent index-writer spawn through it. Otherwise
-            // fall back to the ambient runtime (the calling runtime is the
-            // only one available when `runtime_handle.is_none()`, and the
-            // ambient resolves to it).
-            let handle = match &runtime_handle {
-                Some(h) => h.spawn(task),
-                None => tokio::spawn(task),
-            };
-            *self.index_writer_handle.lock().await = Some(handle);
-        }
-
-        let state = Arc::clone(&self.state);
-        let artifact_load = tokio::task::spawn_blocking(move || {
-            if let Err(e) = state.artifact_store.load_from_disk() {
-                tracing::warn!("embedded artifact index load failed, continuing empty: {e}");
-            }
-            let entries = state.artifact_store.load_all();
-            let count = entries.len();
-            for (key, meta) in entries {
-                state
-                    .artifacts
-                    .insert(key, CachedArtifact::from_index(meta));
-            }
-            state.artifacts_loaded.store(true, Ordering::Release);
-            state.artifact_store_loaded.store(true, Ordering::Release);
-            count
-        })
-        .await
-        .unwrap_or(0);
-        if artifact_load > 0 {
-            tracing::info!(loaded = artifact_load, "embedded artifact index restored");
-        }
-
-        let metadata_state = Arc::clone(&self.state);
-        let metadata_path = self.state.metadata_path.clone();
-        let _ = tokio::task::spawn_blocking(move || {
-            match crate::fscache::MetadataCache::load_from_disk(metadata_path.as_path()) {
-                Ok(loaded) => metadata_state.cache_system.metadata().merge_from(loaded),
-                Err(e) => tracing::warn!(
-                    path = %metadata_path.display(),
-                    "failed to load embedded metadata cache, starting empty: {e}"
-                ),
-            }
-            metadata_state
-                .metadata_cache_loaded
-                .store(true, Ordering::Release);
-        })
-        .await;
-
-        let compiler_state = Arc::clone(&self.state);
-        let compiler_hash_cache_path = self.state.compiler_hash_cache_path.clone();
-        let _ = tokio::task::spawn_blocking(move || {
-            match CompilerHashCache::load_from_disk(compiler_hash_cache_path.as_path()) {
-                Ok(loaded) => compiler_state.compiler_hash_cache.merge_from(loaded),
-                Err(e) => tracing::warn!(
-                    path = %compiler_hash_cache_path.display(),
-                    "failed to load embedded compiler hash cache, starting empty: {e}"
-                ),
-            }
-            compiler_state
-                .compiler_hash_cache_loaded
-                .store(true, Ordering::Release);
-        })
-        .await;
-
-        let includes_state = Arc::clone(&self.state);
-        let system_includes_cache_path = self.state.system_includes_cache_path.clone();
-        let _ = tokio::task::spawn_blocking(move || {
-            match crate::depgraph::SystemIncludeCache::load_from_disk(
-                system_includes_cache_path.as_path(),
-            ) {
-                Ok(loaded) => {
-                    let mut live = includes_state.system_includes.blocking_lock();
-                    live.merge_from(loaded);
-                }
-                Err(e) => tracing::warn!(
-                    path = %system_includes_cache_path.display(),
-                    "failed to load embedded system include cache, starting empty: {e}"
-                ),
-            }
-            includes_state
-                .system_includes_loaded
-                .store(true, Ordering::Release);
-        })
-        .await;
-
-        let depgraph_path = embedded_depgraph_file_path(&self.state);
-        let state = Arc::clone(&self.state);
-        let _ = tokio::task::spawn_blocking(move || {
-            let outcome = crate::depgraph::classify_load(depgraph_path.as_path());
-            let warning = outcome.warning(depgraph_path.as_path());
-            if let Some(graph) = outcome.into_graph() {
-                state.dep_graph.store(Arc::new(graph));
-                state.dep_graph_persisted.store(true, Ordering::Release);
-            }
-            if let Some(warning) = warning {
-                let mut guard = state
-                    .depgraph_load_warning
-                    .lock()
-                    .unwrap_or_else(|poisoned| poisoned.into_inner());
-                *guard = Some(warning);
-            }
-            state.dep_graph_load_complete.store(true, Ordering::Release);
-            state.dep_graph_load_notify.notify_waiters();
-        })
-        .await;
-    }
-
-    pub(crate) async fn compile(
-        &self,
-        request: EmbeddedCompileRequest,
-    ) -> Result<EmbeddedCompileResult, String> {
-        self.state
-            .last_activity
-            .store(now_secs(), Ordering::Relaxed);
-        // zccache#940: per-compile id for the diagnostic trace. The
-        // embedded daemon does not yet surface an audit id through
-        // EmbeddedCompileRequest, so we generate a monotonic per-process
-        // counter here. Hosts that already track their own per-compile
-        // id (soldr's `c<N>` scheme) get a parallel namespace; the two
-        // sides can be cross-correlated by timestamp.
-        let compile_id = next_inner_compile_id();
-        let total = std::time::Instant::now();
-        // soldr#1286: capture journal metadata BEFORE the handler consumes
-        // the request so embedded compiles land in compile_journal.jsonl
-        // exactly like daemon-IPC compiles do (connection.rs journal block).
-        // Without this the embedded backend — the only compile path for
-        // soldr since zccache became an embedded service — was invisible
-        // to hit/miss telemetry: `zccache analyze`, dashboards, and
-        // post-mortem scripts saw zero rustc records.
-        let journal_ctx = JournalContext {
-            compiler: request.compiler.to_string_lossy().into_owned(),
-            args: request.args.clone(),
-            cwd: request.cwd.to_string_lossy().into_owned(),
-            env: request.env.clone(),
-            session_id: None,
-        };
-        let response = handle_compile_ephemeral(
-            &self.state,
-            std::process::id(),
-            &request.cwd,
-            &request.compiler,
-            &request.args,
-            &request.cwd,
-            request.env,
-            request.stdin,
-        )
-        .await;
-        crate::compile_trace::record(
-            "embedded_daemon_compile",
-            total.elapsed().as_micros() as u64,
-            &compile_id,
-        );
-        // Journal the outcome (hit/miss/error + miss_reason) on the same
-        // background-thread writer the daemon path uses. `log` never
-        // blocks, so the embedded hot path pays only the context capture
-        // above plus serde serialization — parity with the IPC path's
-        // accepted cost (issue #459).
-        if let Some((outcome, exit_code, default_reason)) = extract_outcome(&response) {
-            let miss_reason =
-                super::connection::compile_miss_reason(&journal_ctx, outcome, default_reason);
-            let entry = JournalEntry::new(
-                journal_ctx,
-                outcome,
-                exit_code,
-                total.elapsed().as_nanos(),
-                miss_reason,
-            );
-            self.state.journal.log(&entry, None);
-        }
-        match response {
-            Response::CompileResult {
-                exit_code,
-                stdout,
-                stderr,
-                cached,
-            } => {
-                crate::compile_trace::record(
-                    if cached {
-                        "embedded_outcome_cached"
-                    } else {
-                        "embedded_outcome_miss"
-                    },
-                    0,
-                    &compile_id,
-                );
-                Ok(EmbeddedCompileResult {
-                    exit_code,
-                    stdout,
-                    stderr,
-                    cached,
-                })
-            }
-            Response::Error { message } => {
-                crate::compile_trace::record("embedded_outcome_error", 0, &compile_id);
-                Err(message)
-            }
-            other => Err(format!("unexpected embedded compile response: {other:?}")),
-        }
-    }
-
-    pub(crate) async fn stats(&self) -> EmbeddedStatsSnapshot {
-        EmbeddedStatsSnapshot {
-            status: status_snapshot(&self.state).await,
-            phase_profile: self.state.profiler.totals_snapshot().into(),
-        }
-    }
-
-    pub(crate) async fn flush(&self) -> EmbeddedFlushReport {
-        let mut index_writer_handle = self.index_writer_handle.lock().await;
-        flush_embedded_state(&self.state, &mut index_writer_handle, false).await
-    }
-
-    pub(crate) async fn shutdown(&self) -> EmbeddedFlushReport {
-        self.state.shutdown_requested.store(true, Ordering::Release);
-        self.state.index_writer_shutdown.notify_waiters();
-        let mut index_writer_handle = self.index_writer_handle.lock().await;
-        let report = flush_embedded_state(&self.state, &mut index_writer_handle, true).await;
-        let _ = std::fs::remove_dir_all(&self.state.depfile_tmpdir);
-        report
-    }
-}
-
-async fn status_snapshot(state: &SharedState) -> crate::protocol::DaemonStatus {
-    let snap = state.stats.snapshot();
-    let dg = state.dep_graph.load().stats();
-    let artifact_count = state.artifacts.len() as u64;
-    let cache_size_bytes: u64 = state
-        .artifacts
-        .iter()
-        .map(|entry| entry.value().meta.total_size)
-        .sum();
-    let metadata_entries = state.cache_system.metadata().len() as u64;
-    let private_daemon = state.private_daemon.snapshot().await;
-    crate::protocol::DaemonStatus {
-        version: crate::core::VERSION.to_string(),
-        daemon_namespace: state.daemon_namespace.clone(),
-        endpoint: state.endpoint.clone(),
-        private_daemon,
-        artifact_count,
-        cache_size_bytes,
-        metadata_entries,
-        uptime_secs: now_secs().saturating_sub(state.start_time),
-        cache_hits: snap.hits,
-        cache_misses: snap.misses,
-        total_compilations: snap.compilations,
-        non_cacheable: snap.non_cacheable,
-        compile_errors: snap.compile_errors,
-        compile_errors_cached: snap.compile_errors_cached,
-        time_saved_ms: snap.time_saved_ms(),
-        total_links: snap.link_total,
-        link_hits: snap.link_hits,
-        link_misses: snap.link_misses,
-        link_non_cacheable: snap.link_non_cacheable,
-        dep_graph_contexts: dg.context_count as u64,
-        dep_graph_files: dg.file_count as u64,
-        sessions_total: snap.sessions_total,
-        sessions_active: state.sessions.active_count() as u64,
-        cache_dir: state.cache_dir.clone(),
-        dep_graph_version: crate::depgraph::DEPGRAPH_VERSION,
-        dep_graph_disk_size: embedded_depgraph_file_path(state)
-            .metadata()
-            .map(|m| m.len())
-            .unwrap_or(0),
-        dep_graph_persisted: state.dep_graph_persisted.load(Ordering::Acquire),
-    }
-}
-
-async fn flush_embedded_state(
-    state: &Arc<SharedState>,
-    index_writer_handle: &mut Option<tokio::task::JoinHandle<()>>,
-    shutdown_writer: bool,
-) -> EmbeddedFlushReport {
-    let pending_writes_drained = pending_writes::await_all(
-        &state.pending_cache_writes,
-        std::time::Duration::from_secs(30),
-    )
-    .await;
-
-    if shutdown_writer {
-        state.index_writer_shutdown.notify_waiters();
-        if let Some(handle) = index_writer_handle.as_mut() {
-            if !handle.is_finished() {
-                let _ = tokio::time::timeout(std::time::Duration::from_secs(2), handle).await;
-            }
-        }
-    }
-
-    let artifact_entries = state.artifact_store.len() as u64;
-    let _ = Arc::clone(&state.artifact_store).flush_async().await;
-
-    let dg = state.dep_graph.load_full();
-    let depgraph_path = embedded_depgraph_file_path(state);
-    let depgraph_state = Arc::clone(state);
-    let _ = tokio::task::spawn_blocking(move || {
-        if let Some(parent) = depgraph_path.parent() {
-            std::fs::create_dir_all(parent).ok();
-        }
-        if crate::depgraph::save_to_file(&dg, depgraph_path.as_path()).is_ok() {
-            depgraph_state
-                .dep_graph_persisted
-                .store(true, Ordering::Release);
-        }
-    })
-    .await;
-
-    let metadata_entries = state.cache_system.metadata().len() as u64;
-    if state.metadata_cache_loaded.load(Ordering::Acquire) {
-        let metadata_state = Arc::clone(state);
-        let metadata_path = state.metadata_path.clone();
-        let _ = tokio::task::spawn_blocking(move || {
-            metadata_state
-                .cache_system
-                .metadata()
-                .save_to_disk(metadata_path.as_path())
-        })
-        .await;
-    }
-
-    if state.compiler_hash_cache_loaded.load(Ordering::Acquire) {
-        let compiler_state = Arc::clone(state);
-        let compiler_hash_cache_path = state.compiler_hash_cache_path.clone();
-        let _ = tokio::task::spawn_blocking(move || {
-            compiler_state
-                .compiler_hash_cache
-                .save_to_disk(compiler_hash_cache_path.as_path())
-        })
-        .await;
-    }
-
-    if state.system_includes_loaded.load(Ordering::Acquire) {
-        let includes = {
-            let includes = state.system_includes.lock().await;
-            includes.clone()
-        };
-        let system_includes_cache_path = state.system_includes_cache_path.clone();
-        let _ = tokio::task::spawn_blocking(move || {
-            includes.save_to_disk(system_includes_cache_path.as_path())
-        })
-        .await;
-    }
-
-    EmbeddedFlushReport {
-        pending_writes_drained,
-        artifact_entries,
-        metadata_entries,
-    }
-}
-
-fn embedded_depgraph_file_path(state: &SharedState) -> crate::core::NormalizedPath {
-    state.cache_dir.join("depgraph").join("depgraph.bin")
-}
 
 impl DaemonServer {
     /// Get a handle to signal shutdown.
@@ -688,85 +301,6 @@ impl DaemonServer {
         *guard = Some(warning);
     }
 
-    /// Hand out an owned handle that can install a loaded depgraph (and
-    /// optional warning) from any thread, at any time during the daemon's
-    /// lifetime. This is the #640 seam for the deferred-depgraph-load —
-    /// the background `spawn_blocking` loader holds a `DepGraphSetter` and
-    /// calls `install()` when the disk load completes; meanwhile
-    /// `server.run()` has been accepting connections from t ≈ 0.
-    pub fn dep_graph_setter(&self) -> DepGraphSetter {
-        DepGraphSetter {
-            state: Arc::clone(&self.state),
-        }
-    }
-
-    /// Hand out a loader that reads the compiler-hash cache from disk and
-    /// installs it into the running state. Issue #784 — moves the load
-    /// out of `bind_with_cache_dir`'s critical path so the readiness
-    /// lockfile fires before any disk I/O.
-    ///
-    /// Designed to be called once, immediately after `write_lock_file`,
-    /// inside a `tokio::task::spawn_blocking`. The handle captures the
-    /// on-disk snapshot path so the daemon binary doesn't need access
-    /// to the (private) `compiler_hash_cache_path` field.
-    pub fn compiler_hash_cache_loader(&self) -> CompilerHashCacheLoader {
-        CompilerHashCacheLoader {
-            state: Arc::clone(&self.state),
-            path: self.state.compiler_hash_cache_path.clone(),
-        }
-    }
-
-    /// Hand out a loader that reads the on-disk `metadata.bin` snapshot
-    /// and installs it into the running state's `CacheSystem`. Issue
-    /// #784 phase 2b — extends the compiler-hash-cache deferral above
-    /// to the biggest of the four snapshots (the one that scales with
-    /// the cache size).
-    ///
-    /// Designed to be called once, immediately after `write_lock_file`,
-    /// inside a `tokio::task::spawn_blocking`. The handle captures the
-    /// metadata path so the daemon binary doesn't need access to the
-    /// private `metadata_path` field on `SharedState`.
-    pub fn metadata_cache_loader(&self) -> MetadataCacheLoader {
-        MetadataCacheLoader {
-            state: Arc::clone(&self.state),
-            path: self.state.metadata_path.clone(),
-        }
-    }
-
-    /// Hand out a loader that reads the on-disk system-includes snapshot
-    /// and merges it into the live `Mutex<SystemIncludeCache>` on the
-    /// state. Issue #784 phase 2c.
-    ///
-    /// Designed to be called once, immediately after `write_lock_file`,
-    /// inside a `tokio::task::spawn_blocking`. The loader uses
-    /// `blocking_lock()` on the tokio mutex so the brief merge fits
-    /// inside the blocking thread (same shape as
-    /// `set_depgraph_load_warning`).
-    pub fn system_includes_loader(&self) -> SystemIncludesLoader {
-        SystemIncludesLoader {
-            state: Arc::clone(&self.state),
-            path: self.state.system_includes_cache_path.clone(),
-        }
-    }
-
-    /// Hand out a loader that reads the on-disk `index.bin` blob and
-    /// merges its entries into the live `ArtifactStore`. Issue #784
-    /// phase 2d — last of the four #784 deferrals.
-    ///
-    /// Designed to be called once, immediately after `write_lock_file`,
-    /// inside a `tokio::task::spawn_blocking`. The handle holds an
-    /// `Arc<ArtifactStore>` clone so the merge writes hit the same
-    /// `DashMap` request handlers read. If a lookup races ahead of
-    /// this background load,
-    /// `util::lookup_artifact_with_disk_fallback` invokes
-    /// [`ArtifactStore::load_from_disk`] synchronously on the spot —
-    /// idempotent because both call sites do equivalent inserts.
-    pub fn artifact_store_loader(&self) -> ArtifactStoreLoader {
-        ArtifactStoreLoader {
-            state: Arc::clone(&self.state),
-            store: Arc::clone(&self.state.artifact_store),
-        }
-    }
 
     /// Get a snapshot of the phase profiler (for benchmarks).
     #[must_use]
@@ -854,236 +388,6 @@ impl DaemonServer {
     }
 }
 
-/// Owned handle to install a loaded depgraph on a running [`DaemonServer`].
-///
-/// Holds a clone of the daemon's `Arc<SharedState>`. The setter survives
-/// across `tokio::task::spawn_blocking` boundaries (it is `Send + Sync`),
-/// so `bin/zccache-daemon::run_server` can hand one off to the background
-/// loader BEFORE calling `server.run()`. When the disk load finishes, the
-/// loader calls [`Self::install`] and the daemon's hot-path readers
-/// (`state.dep_graph.load()` at every compile request) atomically pick up
-/// the new graph on their next `.load()`.
-///
-/// Issue #640.
-pub struct DepGraphSetter {
-    state: Arc<SharedState>,
-}
-
-impl DepGraphSetter {
-    /// Atomically install a loaded depgraph (and optional warning).
-    ///
-    /// - `graph = Some(g)` swaps the daemon's empty default with the
-    ///   loaded graph and marks the on-disk snapshot as persisted (so
-    ///   the next clean shutdown doesn't re-save the empty default
-    ///   over the real graph).
-    /// - `graph = None` leaves the empty default in place. Use this
-    ///   for the `Missing` / corrupt-load fallback so the warning
-    ///   still routes into the per-session log via `warning`.
-    /// - `warning` is mirrored into `SessionStart`'s per-session log
-    ///   if `Some` (issue #320).
-    pub fn install(self, graph: Option<crate::depgraph::DepGraph>, warning: Option<String>) {
-        if let Some(graph) = graph {
-            self.state.dep_graph.store(Arc::new(graph));
-            self.state
-                .dep_graph_persisted
-                .store(true, Ordering::Release);
-        }
-        if let Some(warning) = warning {
-            let mut guard = self
-                .state
-                .depgraph_load_warning
-                .lock()
-                .unwrap_or_else(|poisoned| poisoned.into_inner());
-            *guard = Some(warning);
-        }
-        self.state
-            .dep_graph_load_complete
-            .store(true, Ordering::Release);
-        self.state.dep_graph_load_notify.notify_waiters();
-    }
-}
-
-/// Owned handle that loads the on-disk compiler-hash-cache snapshot and
-/// merges it into the running daemon's state.
-///
-/// Issue #784. Mirrors [`DepGraphSetter`]'s role for the depgraph: the
-/// daemon binary holds one across a `spawn_blocking` boundary and calls
-/// [`Self::load_and_install`] once after the readiness lockfile is
-/// written. The merge uses [`CompilerHashCache::merge_from`], which is
-/// `&self`, so concurrent `get_or_hash_with` readers either see no entry
-/// (cold-hash path) or a loaded entry (stat-verify guards correctness).
-pub struct CompilerHashCacheLoader {
-    state: Arc<SharedState>,
-    path: crate::core::NormalizedPath,
-}
-
-/// Owned handle that loads the on-disk system-includes snapshot and
-/// merges it into the running daemon's `Mutex<SystemIncludeCache>`.
-///
-/// Issue #784 phase 2c. Same shape as [`CompilerHashCacheLoader`] but
-/// uses `blocking_lock()` to acquire the tokio mutex briefly during
-/// the merge (the `SystemIncludeCache` itself uses `HashMap` with
-/// `&mut self` mutations, so the mutex stays on the live field).
-pub struct SystemIncludesLoader {
-    state: Arc<SharedState>,
-    path: crate::core::NormalizedPath,
-}
-
-/// Owned handle that reads the on-disk artifact-index blob and merges
-/// its entries into the running daemon's `ArtifactStore`.
-///
-/// Issue #784 phase 2d. Holds an `Arc<ArtifactStore>` directly so the
-/// merge writes hit the same `DashMap` the request handlers read — no
-/// swap needed (the store was constructed empty at bind time).
-///
-/// Coexists safely with the on-demand `load_from_disk` invocation
-/// inside `util::lookup_artifact_with_disk_fallback`: both call sites
-/// perform identical `DashMap::insert` operations, so a race between
-/// them produces the same converged state. The `artifact_store_loaded`
-/// flag merely prevents redundant disk reads — it is not a load-once
-/// gate.
-pub struct ArtifactStoreLoader {
-    state: Arc<SharedState>,
-    store: Arc<ArtifactStore>,
-}
-
-impl ArtifactStoreLoader {
-    /// Read the on-disk index blob (if any) and insert its entries
-    /// into the live store via [`ArtifactStore::load_from_disk`].
-    ///
-    /// I/O errors are logged at WARN and treated as "empty" — the
-    /// daemon stays running with an empty in-memory index, and the
-    /// next periodic flush will rewrite the file from whatever
-    /// request-handler inserts have landed since. After the load,
-    /// `artifact_store_loaded` is set so subsequent on-demand calls
-    /// from `lookup_artifact_with_disk_fallback` short-circuit.
-    pub fn load_and_install(self) {
-        if let Err(e) = self.store.load_from_disk() {
-            tracing::warn!("artifact index load failed, continuing with empty store: {e}");
-        }
-        self.state
-            .artifact_store_loaded
-            .store(true, Ordering::Release);
-    }
-}
-
-impl SystemIncludesLoader {
-    /// Load the on-disk snapshot (if any) and merge it into the live
-    /// cache.
-    ///
-    /// A missing or corrupt snapshot is logged at WARN and the live
-    /// cache is left empty (the stat-verify safety net in
-    /// `SystemIncludeCache::get` / `get_or_discover` keeps correctness
-    /// either way). After the merge — successful or empty —
-    /// `system_includes_loaded` is set so `run.rs`'s shutdown path
-    /// knows the in-memory state is canonical and safe to save.
-    pub fn load_and_install(self) {
-        match crate::depgraph::SystemIncludeCache::load_from_disk(self.path.as_path()) {
-            Ok(loaded) => {
-                let loaded_len = loaded.len();
-                {
-                    let mut live = self.state.system_includes.blocking_lock();
-                    live.merge_from(loaded);
-                }
-                if loaded_len > 0 {
-                    tracing::info!(
-                        loaded = loaded_len,
-                        path = %self.path.display(),
-                        "system include cache restored from disk (background)"
-                    );
-                }
-            }
-            Err(e) => {
-                tracing::warn!(
-                    path = %self.path.display(),
-                    "failed to load system include cache, starting empty: {e}"
-                );
-            }
-        }
-        self.state
-            .system_includes_loaded
-            .store(true, Ordering::Release);
-    }
-}
-
-impl CompilerHashCacheLoader {
-    /// Load the on-disk compiler-hash-cache snapshot (if any) and merge
-    /// it into the live state.
-    ///
-    /// A missing or corrupt snapshot is logged at WARN and the live
-    /// cache is left empty (the stat-verify safety net in
-    /// `get_or_hash_with` keeps correctness either way). After the
-    /// merge — successful or empty — `compiler_hash_cache_loaded` is
-    /// set so `run.rs`'s shutdown path knows the in-memory state is
-    /// canonical and safe to save.
-    pub fn load_and_install(self) {
-        match CompilerHashCache::load_from_disk(self.path.as_path()) {
-            Ok(loaded) => {
-                self.state.compiler_hash_cache.merge_from(loaded);
-            }
-            Err(e) => {
-                tracing::warn!(
-                    path = %self.path.display(),
-                    "failed to load compiler hash cache, starting empty: {e}"
-                );
-            }
-        }
-        self.state
-            .compiler_hash_cache_loaded
-            .store(true, Ordering::Release);
-    }
-}
-
-/// Owned handle that loads the on-disk `metadata.bin` snapshot and
-/// merges it into the running daemon's `CacheSystem`.
-///
-/// Issue #784 phase 2b. Same shape as [`CompilerHashCacheLoader`] —
-/// the daemon binary holds one across a `spawn_blocking` boundary and
-/// calls [`Self::load_and_install`] once after the readiness lockfile
-/// is written. The merge uses [`crate::fscache::MetadataCache::merge_from`],
-/// which is `&self`, so concurrent `get_cached_hash_if_stat_valid`
-/// readers either see no entry (cold-path miss — re-stat + re-hash)
-/// or a loaded entry (stat-verify guards correctness).
-pub struct MetadataCacheLoader {
-    state: Arc<SharedState>,
-    path: crate::core::NormalizedPath,
-}
-
-impl MetadataCacheLoader {
-    /// Load the on-disk `metadata.bin` snapshot (if any) and merge it
-    /// into the live state.
-    ///
-    /// A missing or corrupt snapshot is logged at WARN and the live
-    /// cache is left empty (the stat-verify safety net in
-    /// `get_cached_hash_if_stat_valid` keeps correctness either way).
-    /// After the merge — successful or empty — `metadata_cache_loaded`
-    /// is set so `run.rs`'s shutdown path knows the in-memory state is
-    /// canonical and safe to save.
-    pub fn load_and_install(self) {
-        match crate::fscache::MetadataCache::load_from_disk(self.path.as_path()) {
-            Ok(loaded) => {
-                let loaded_len = loaded.len();
-                self.state.cache_system.metadata().merge_from(loaded);
-                if loaded_len > 0 {
-                    tracing::info!(
-                        loaded = loaded_len,
-                        path = %self.path.display(),
-                        "metadata cache restored from disk (background)"
-                    );
-                }
-            }
-            Err(e) => {
-                tracing::warn!(
-                    path = %self.path.display(),
-                    "failed to load metadata cache, starting empty: {e}"
-                );
-            }
-        }
-        self.state
-            .metadata_cache_loaded
-            .store(true, Ordering::Release);
-    }
-}
 
 #[cfg(test)]
 mod tests {
